src/remittance/requirements.txt
attrs==19.3.0
boto3==1.12.39
botocore==1.15.39
instabot==0.117.0
pandas==1.0.3
Pillow==7.1.1
selenium==3.141.0
